pyproject.toml
[project]
name = "claude-assistant"
version = "0.2.0"
description = "Voice-enabled Claude Code assistant with programmable automation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Myra", email = "myra@example.com" }
]

dependencies = [
    "pvcheetah>=2.0.0",
    "pvporcupine>=3.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "markdown>=3.4.0",
    "tqdm>=4.66.0,<4.67.0",
    "anthropic>=0.40.0",
    "webrtcvad>=2.0.10",
    "pyyaml>=6.0",
    "pyrubberband>=0.3.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
all = [
    "claude-assistant[mcp,dev]",
]

[project.scripts]
claude-assistant = "voice_assistant.main:main"
claude-voice = "voice_assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_assistant", "src/mcp"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
